requirements.txt
alembic==1.0.9
asn1crypto==0.24.0
astroid==2.0.4
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
cov-core==1.15.0
coverage==4.4.1
cryptography==2.3.1
cycler==0.10.0
Cython==0.29
freezegun==0.3.10
idna==2.6
int-date==0.1.8
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.25.0
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
more-itertools==4.3.0
nose2==0.6.5
nose2[coverage-plugin]
numba==0.35.0
numpy==1.13.3
pandas==0.20.3
py==1.7.0
pycparser==2.19
pylint==2.1.1
PyMySQL==0.7.11
pyOpenSSL==18.0.0
pyparsing==2.2.2
pytest==3.2.2
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
requests==2.18.4
scipy==1.2.1
six==1.10.0
SQLAlchemy==1.1.18
stockstats==0.2.0
tulipy==0.3.1
typed-ast==1.1.0
urllib3==1.22
wrapt==1.10.11
unidecode
PyYAML==5.1